Time stamp to DAQ changed to GPS seconds unsigned int instead of double. Controller.c updated at same time to send GPS second info.
git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@2084 6dcd42c9-f523-4c6d-aada-af552506706e

+++ src/include/drv/fb.h
@@ -37,7 +37,7 @@ int cdsDaqNetDaqSend(   int dcuId,
 #endif
 
 extern int cdsNetStatus;
-extern double cycle_gps_time;
+extern unsigned int cycle_gps_time;
 extern unsigned int cycle_gps_ns;
 
 #endif
